requirements.txt
tensorflow-cpu==2.13.0
keras==2.13.1
numpy
pillow
uvicorn
fastapi
python-multipart